From vendor tag: 'LA.VENDOR.15.4.0.r1-20300-pakala.QSSI16.0'

* https://git.codelinaro.org/clo/la/platform/vendor/qcom-opensource/wlan/qca-wifi-host-cmn:
  qcacmn: Add support for Kernel LTS 6.16 upgrade
  qcacmn: Fix issue of napi_struct being deleted multiple times
  qcacmn: Handling memory leaks in dummy netdev
  qcacmn: Add api to get ho fail disconnect command pending
  qcacmn: Validate num_switch_trigger_reason with num_tlv
  qcacmn: Keep counter atomicity while logging
  qcacmn: Fix multi-link setup failure issue
  qcacmn: Fix rx_bufs_reaped bounds and tx_status_info addr2 cast
  qcacmn: Initialize beacon rssi history report vdev with invalid
  qcacmn: Enable feature flag & stats id
  qcacmn: Add new INI enable_bcn_rssi_history_report
  qcacmn: Extract beacon RSSI history info from firmware
  qcacmn: Fix out-of-bounds access in wlan_crypto_is_mmie_valid()
  qcacmn: Dump DCS stats in case of negative TSF value
  qcacmn: Fix memory deallocation more than once
  qcacmn: Fix multiple out-of-bounds memory access
  qcacmn: Fix OOB write in extract_additional_cli_rules_meta_info()

+++ qcom/opensource/wlan/qca-wifi-host-cmn/hal/wifi3.0/be/hal_be_generic_api.h
@@ -1635,8 +1635,8 @@ hal_txmon_status_parse_tlv_generic_be(hal_soc_handle_t hal_soc_hdl,
 				HAL_TX_MON_TLV_GET(tx_tlv,
 						   RX_RESPONSE_REQUIRED_INFO,
 						   ADDR1_31_0);
-		*(uint32_t *)&tx_status_info->addr2[4] =
-				HAL_TX_MON_TLV_GET(tx_tlv,
+		*(uint16_t *)&tx_status_info->addr2[4] =
+				(uint16_t)HAL_TX_MON_TLV_GET(tx_tlv,
 						   RX_RESPONSE_REQUIRED_INFO,
 						   ADDR1_47_32);
 		*(uint32_t *)&tx_status_info->addr1[0] =
